fix(chat-core): isolate contracts-free CheckpointStore port surface for api typecheck
