+++ packages/chat-core/src/index.ts
@@ -1,2 +1,3 @@
+export * from './checkpoint-port.js';
 export * from './ports.js';
 export * from './types.js';
