+++ packages/chat-core/src/ports.ts
@@ -14,44 +14,13 @@ export type {
 } from '@sentropic/contracts';
 export type { EventEnvelope, StreamEvent } from '@sentropic/events';
 
-/**
- * Per SPEC §12 — generic CheckpointStore with strategy adapters.
- * Single port owns versioning, listing, tagging, forking, optional watch.
- * Strategy concerns (cadence, redaction) live in the adapter, not the port.
- */
-export type CheckpointMeta = {
-  key: string;
-  version: number;
-  tags?: ReadonlyArray<string>;
-  createdAt: string;
-  updatedAt: string;
-};
-
-export type SaveResult = {
-  version: number;
-  success: boolean;
-  reason?: 'VersionMismatch';
-};
-
-export interface CheckpointStore<T> {
-  load(key: string): Promise<{ state: T; version: number } | null>;
-  save(
-    key: string,
-    state: T,
-    expectedVersion?: number,
-  ): Promise<SaveResult>;
-  list(
-    prefix?: string,
-    limit?: number,
-  ): Promise<ReadonlyArray<CheckpointMeta>>;
-  delete(key: string): Promise<void>;
-  tag(key: string, label: string): Promise<void>;
-  fork(sourceKey: string, targetKey: string): Promise<void>;
-  watch?(
-    key: string,
-    callback: (state: T, version: number) => void,
-  ): () => void;
-}
+// Re-export the contracts-free CheckpointStore surface so existing
+// `from '@sentropic/chat-core'` consumers keep working unchanged.
+export type {
+  CheckpointMeta,
+  CheckpointStore,
+  SaveResult,
+} from './checkpoint-port.js';
 
 /**
  * Per SPEC §5 — chat-core ports.
